update: Update dependencies and codegen-units***
This commit updates the dependencies in Cargo.lock and increases the codegen-units value in Cargo.toml.
