+++ Cargo.toml
@@ -10,4 +10,4 @@ git2 = "0.18.2"
 
 [profile.release]
 lto = true
-debug = 0
+codegen-units = 1
