+++ src/main.rs
@@ -1,21 +1,7 @@
-use git2::{Error as GitError, Repository};
-use std::io::Error as IoError;
-use std::{env, fs, process};
+use prepare_commit_msg::{get_branch_name, update_commit_message};
+use std::{env, process};
 
 const BRANCH_NAMES: [&str; 2] = ["main", "master"];
-
-fn get_branch_name() -> Result<String, GitError> {
-    let repo = Repository::open(".")?;
-    let head = repo.head()?;
-    Ok(head.shorthand().unwrap_or("").to_string())
-}
-
-fn update_commit_message(commit_msg_file: &str, branch_name: &str) -> Result<(), IoError> {
-    let commit_msg = fs::read_to_string(commit_msg_file)?;
-    let new_commit_msg = format!("{}: {}", branch_name, commit_msg);
-    fs::write(commit_msg_file, new_commit_msg)?;
-    Ok(())
-}
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -44,23 +30,4 @@ fn main() {
         Ok(()) => println!("Commit message updated with branch name."),
         Err(err) => eprintln!("Failed to update commit message: {}", err),
     }
-}
-
-#[cfg(test)]
-#[test]
-fn test_get_branch_name() {
-    let branch_name = get_branch_name().unwrap();
-    assert!(!branch_name.is_empty());
-}
-
-#[test]
-fn test_update_commit_message() {
-    let commit_msg_file = "test_commit_msg.txt";
-    let branch_name = "test-branch";
-    let commit_msg = "Test commit message";
-    fs::write(commit_msg_file, commit_msg).unwrap();
-    update_commit_message(commit_msg_file, branch_name).unwrap();
-    let new_commit_msg = fs::read_to_string(commit_msg_file).unwrap();
-    assert_eq!(new_commit_msg, format!("{}: {}", branch_name, commit_msg));
-    fs::remove_file(commit_msg_file).unwrap();
 }
